backend/requirements.txt
Flask==3.0.3
pydantic==2.8.2
handcalcs==1.11.0
supabase==2.10.0
python-dotenv==1.0.1
